🍆 feat: calculator's compute method returns decimal

+++ src/infrastructure/calculators/base.py
@@ -5,6 +5,7 @@ import logging
 
 from abc import ABC, abstractmethod
 from dataclasses import dataclass
+from decimal import Decimal
 
 from infrastructure.common.base import InfrastructureBase
 
@@ -14,7 +15,7 @@ logger = logging.getLogger("app")
 @dataclass
 class CalculatorBase(InfrastructureBase, ABC):
     @abstractmethod
-    async def compute(self, expression: str) -> float:
+    async def compute(self, expression: str) -> Decimal:
         """-------------------------------------------------------------------------
         Compute the expression and return the result.
         -------------------------------------------------------------------------"""
